refactoring new-review into server comp/client comp split
